# faster `BigFloat` construction for more types