--- a/base/mpfr.jl
+++ b/base/mpfr.jl
@@ -306,7 +306,7 @@ function BigFloat(x::Base.BitInteger64, r::MPFRRoundingMode=ROUNDING_MODE[]; pre
         return z
     else
         # If Clong is 32 bit, MPFR does not have a method for (U)Int64 on Windows
-        return sizeof(x) <= sizeof(Clong) ? 
+        return sizeof(x) <= sizeof(Clong) ?
             BigFloat_mpfr_fallback(x, r; precision)::BigFloat :
             BigFloat(BigInt(x), r; precision)
     end