This pull request addresses #50940 and is a follow-up to #47546.

The changes add Julia construction of `BigFloat`s for some inputs:
- When the Base.BitInteger64 or Base.IEEEFloat when precision is enough to perfectly represent the input as a `BigFloat`
- When the denominator of the `Rational` is a power of 2.

```julia
julia> @btime BigFloat(128)
  14.320 ns (2 allocations: 104 bytes)
128.0

julia> @btime Base.MPFR.BigFloat_mpfr_fallback(128)
  25.518 ns (2 allocations: 104 bytes)
128.0

julia> @btime BigFloat(128.0f0)
  17.952 ns (2 allocations: 104 bytes)
128.0

julia> @btime Base.MPFR.BigFloat_mpfr_fallback(128.0f0)
  35.667 ns (2 allocations: 104 bytes)
128.0
```