@vtjnash 2024-02-06
```
numbers                                          (12) |         failed at 2024-01-13T01:37:41.184
2024-01-12 20:37:41 EST	Test Failed at /cache/build/tester-amdci4-15/julialang/julia-master/julia-770900211a/share/julia/test/numbers.jl:2929
2024-01-12 20:37:41 EST	  Expression: all((m->begin
2024-01-12 20:37:41 EST	            #= /cache/build/tester-amdci4-15/julialang/julia-master/julia-770900211a/share/julia/test/numbers.jl:2929 =#
2024-01-12 20:37:41 EST	            m.file === Symbol("deprecated.jl")
2024-01-12 20:37:41 EST	        end), (collect(methods(T)))[findall((R->(#= /cache/build/tester-amdci4-15/julialang/julia-master/julia-770900211a/share/julia/test/numbers.jl:2930 =#
2024-01-12 20:37:41 EST	                    !(R <: T))), Base.return_types(T))])
2024-01-12 20:37:41 EST	
```
Seems like there may be an inference issue. Maybe @oscardssmith can help get this finished though, since it seems a good performance improvement?

Also, might want to sprinkle around some `GC.@preserve` near the unsafe load/store calls just to be careful.