@hycakir 2024-02-28
Where would I need `GC.@preserve`? It seems all calls to `unsafe_store` makes explicit use of the relevant objects. GC should not collect these objects, no?